+++ CHANGELOG.md
@@ -8,6 +8,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Some relative paths didn't behave as expected before.
+
+
 ## [1.0.2] - 2023-04-16
 
 ### Added